move all dashboard routes from the home routes to a new dashboard routes file, update index to reflect changes
