+++ controllers/index.js
@@ -2,10 +2,14 @@ const router = require('express').Router();
 
 const apiRoutes = require('./api');
 const homeRoutes = require('./homeRoutes');
+const dashboardRoutes = require('./dashboardRoutes');
 const profileRoutes = require('./profileRoutes');
+// const connectRoutes = require('./connectRoutes');
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
+router.use('/dashboard', dashboardRoutes);
 router.use('/profile', profileRoutes);
+// router.use('/connect', connectRoutes);
 
 module.exports = router;
